ValidationAPP/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
alembic==1.8.0
    # via -r requirements.in
anyio==3.6.1
    # via starlette
attrs==21.4.0
    # via pytest
black==22.6.0
    # via -r requirements.in
boto3==1.24.29
    # via -r requirements.in
botocore==1.27.29
    # via
    #   boto3
    #   s3transfer
cachetools==5.2.0
    # via premailer
certifi==2022.6.15
    # via requests
cffi==1.15.1
    # via cryptography
cfgv==3.3.1
    # via pre-commit
chardet==5.0.0
    # via emails
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via
    #   black
    #   uvicorn
cryptography==37.0.4
    # via jwt
cssselect==1.1.0
    # via premailer
cssutils==2.4.2
    # via
    #   emails
    #   premailer
distlib==0.3.4
    # via virtualenv
dnspython==2.2.1
    # via email-validator
ecdsa==0.18.0
    # via python-jose
email-validator==1.2.1
    # via -r requirements.in
emails==0.6
    # via -r requirements.in
fastapi==0.78.0
    # via -r requirements.in
filelock==3.7.1
    # via virtualenv
greenlet==1.1.2
    # via sqlalchemy
h11==0.13.0
    # via uvicorn
identify==2.5.1
    # via pre-commit
idna==3.3
    # via
    #   anyio
    #   email-validator
    #   requests
iniconfig==1.1.1
    # via pytest
jinja2==3.1.2
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jwt==1.3.1
    # via -r requirements.in
lxml==4.9.1
    # via
    #   emails
    #   premailer
mako==1.2.1
    # via alembic
mangum==0.15.0
    # via -r requirements.in
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
mypy-extensions==0.4.3
    # via black
nodeenv==1.7.0
    # via pre-commit
packaging==21.3
    # via pytest
pathspec==0.9.0
    # via black
platformdirs==2.5.2
    # via
    #   black
    #   virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==2.20.0
    # via -r requirements.in
premailer==3.10.0
    # via emails
psycopg2-binary==2.9.3
    # via -r requirements.in
py==1.11.0
    # via pytest
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pycryptodome==3.15.0
    # via -r requirements.in
pydantic==1.9.1
    # via
    #   -r requirements.in
    #   fastapi
pyparsing==3.0.9
    # via packaging
pytest==7.1.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   emails
python-dotenv==0.20.0
    # via -r requirements.in
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
pyyaml==6.0
    # via pre-commit
requests==2.28.1
    # via
    #   -r requirements.in
    #   emails
    #   premailer
rsa==4.8
    # via python-jose
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   ecdsa
    #   python-dateutil
    #   python-multipart
    #   virtualenv
sniffio==1.2.0
    # via anyio
sqlalchemy==1.4.39
    # via
    #   -r requirements.in
    #   alembic
starlette==0.19.1
    # via
    #   fastapi
    #   starlette-context
starlette-context==0.3.4
    # via -r requirements.in
tenacity==8.0.1
    # via -r requirements.in
toml==0.10.2
    # via pre-commit
tomli==2.0.1
    # via
    #   black
    #   pytest
typing-extensions==4.3.0
    # via
    #   black
    #   mangum
    #   pydantic
    #   starlette
urllib3==1.26.10
    # via
    #   botocore
    #   requests
uvicorn==0.18.2
    # via -r requirements.in
virtualenv==20.15.1
    # via pre-commit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
